requirements.txt
Django==4.2
psycopg2-binary==2.9.8
pycryptodomex
djangorestframework
argon2-cffi
django-axes
pyotp
qrcode
Pillow
python-decouple
dj-database-url
django-csp